+++ src/main/java/com/devbp/syncspace/domain/dtos/UpdateBookingRequest.java
@@ -10,6 +10,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -25,7 +27,7 @@ public class UpdateBookingRequest {
     private BookingStatus bookingStatus;
 
     @PositiveOrZero
-    private double pricePaid;
+    private BigDecimal pricePaid;
 
     private PaymentStatus paymentStatus;
